+++ src/main/java/com/imsweb/naaccrxml/runtime/NaaccrPatientConverter.java
@@ -275,6 +275,8 @@ public class NaaccrPatientConverter implements Converter {
         if (Boolean.TRUE.equals(_context.getOptions().getTranslateRenamedItemIds())) {
             if (NaaccrFormat.NAACCR_VERSION_180.equals(_context.getDictionary().getNaaccrVersion()))
                 rawId = NaaccrXmlDictionaryUtils.getRenamedLongNaaccr18Ids().getOrDefault(rawId, rawId);
+            if (_context.getOptions().getItemIdsTranslation() != null)
+                rawId = _context.getOptions().getItemIdsTranslation().getOrDefault(rawId, rawId);
         }
 
         if (!_context.getOptions().processItem(rawId))